fixed restructuring
