+++ extension/src-language-server/stpa/diagram/diagram-generator.ts
@@ -21,12 +21,12 @@ import { SLabel, SModelElement, SModelRoot, SNode } from 'sprotty-protocol';
 
 import { Command, Hazard, Model, Node, SystemConstraint, VE, isContext, isHazard, isSystemConstraint, isUCA } from '../../generated/ast';
 import { StpaServices } from '../stpa-module';
-import { getAspect, leafElement } from '../utils';
+import { collectElementsWithSubComps, getAspect, leafElement } from '../utils';
 import { filterModel } from './filtering';
 import { CSEdge, CSNode, ParentNode, STPAEdge, STPANode, STPAPort } from './stpa-interfaces';
 import { CS_EDGE_TYPE, CS_NODE_TYPE, DUMMY_NODE_TYPE, EdgeType, PARENT_TYPE, PortSide, STPAAspect, STPA_EDGE_TYPE, STPA_INTERMEDIATE_EDGE_TYPE, STPA_NODE_TYPE, STPA_PORT_TYPE } from './stpa-model';
 import { StpaSynthesisOptions, labelManagementValue, showLabelsValue } from './synthesis-options';
-import { collectElementsWithSubComps, createUCAContextDescription, getTargets, setLevelOfCSNodes, setLevelsForSTPANodes } from './utils';
+import { createUCAContextDescription, getTargets, setLevelOfCSNodes, setLevelsForSTPANodes } from './utils';
 
 export class StpaDiagramGenerator extends LangiumDiagramGenerator {
